+++ assets/ar/semantics/STS/STSTrack1_GPT4_FewShot.py
@@ -1,13 +1,13 @@
 import os
 
-from llmebench.datasets import STSArSemEval17Track1Dataset
+from llmebench.datasets import SemEval17T1STSDataset
 from llmebench.models import GPTChatCompletionModel
 from llmebench.tasks import STSTrack1Task
 
 
 def config():
     return {
-        "dataset": STSArSemEval17Track1Dataset,
+        "dataset": SemEval17T1STSDataset,
         "dataset_args": {},
         "task": STSTrack1Task,
         "task_args": {},
