+++ assets/ar/sequence_tagging_and_information_extraction/segmentation/Segmentation_GPT35_ZeroShot.py
@@ -2,7 +2,7 @@ import os
 import re
 
 from llmebench.datasets import ArabicSegmentationDataset
-from llmebench.models import GPTModel, RandomGPTModel
+from llmebench.models import GPTModel
 from llmebench.tasks import ArabicSegmentationTask
